Pull information and run a test policy against
Firstly, opted against re-inventing the wheel by querying the APIs manually, as going forward, it will help with rate limiting and other queries we can pass into the functions

A unit test has also been created to simulate the output from the /org/{org} endpoint so we can replicate items uniformly. In the future, proper fixtures should be put in place so we can test unauthenticated responses from GH which will fail a lot of the audit rules due to the sensitive nature

Finally, A single test rego policy is in place just to aide with testing. Future policies will be created in an isolated repository

References BCH-868